Fixed warning in "rdgen"

+++ kernel/utils/rdgen.c
@@ -15,7 +15,7 @@ int main(char argc, char **argv)
 	
 	int nheaders = (argc-2);
 	struct initrd_header headers[64];
-	printf("Ramdisk Header Size: %d\n", sizeof(struct initrd_header));
+	printf("Ramdisk Header Size: %d\n", (int)sizeof(struct initrd_header));
 	unsigned int off = sizeof(struct initrd_header) * 64 + sizeof(int);
 
 	for(int i = 0; i < nheaders; i++)
